+++ Cargo.toml
@@ -1,5 +1,21 @@
+[workspace.package]
+edition = "2024"
+
 [workspace]
 resolver = "3"
 members = [
   "incpa",
+  "incpa-tokio",
 ]
+
+# Note: this serves as a roster of all dependencies w/ versions, not only common dependencies.
+[workspace.dependencies]
+anyhow = "1.0.95"
+derive-new = "0.7.0"
+either = "1.13.0"
+test-case = "3.3.1"
+thiserror = "2.0.9"
+tokio = { version = "1.42.0", default-features = false }
+
+# Local crates:
+incpa = { path = "incpa" }
