+++ incpa-tokio/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "incpa-tokio"
+version = "0.1.0"
+edition.workspace = true
+
+[dependencies]
+incpa.workspace = true
+tokio = { workspace = true, default-features = false, features = ["io-util"] }
