+++ incpa/Cargo.toml
@@ -1,17 +1,14 @@
 [package]
 name = "incpa"
 version = "0.0.1"
-edition = "2021"
-
-[features]
-tokio = ["dep:tokio"]
+edition.workspace = true
 
 [dependencies]
-derive-new = "0.7.0"
-either = "1.13.0"
-thiserror = "2.0.9"
-tokio = { version = "1.42.0", optional = true, default-features = false, features = ["io-util"] }
+derive-new.workspace = true
+either.workspace = true
+thiserror.workspace = true
+tokio = { workspace = true, default-features = false, features = ["io-util"] }
 
 [dev-dependencies]
-anyhow = "1.0.95"
-test-case = "3.3.1"
+anyhow.workspace = true
+test-case.workspace = true
